chore: cleanup matmult files
