+++ jolt-core/src/jolt_onnx/precompiles/mod.rs
@@ -1,4 +1,3 @@
 //! This module implements the sum-check precompiles for ONNX operators.
 pub mod matmult;
-pub mod matmult2;
 pub mod sumcheck_engine;
